SurveyJS survey
This is a survey system created using SurveyJS  JQuery C# and SQL server and database. This system helps one store surveys to the database as a JSON then the JSON is fetched from the sql server and rendered using SurveyJs .The responses are then saved back to the database and can still be edited.